+++ rust/operator-binary/src/main.rs
@@ -29,6 +29,7 @@ use stackable_operator::{
     logging::controller::report_controller_reconciled,
     shared::yaml::SerializeOptions,
     telemetry::Tracing,
+    utils::signal::SignalWatcher,
 };
 
 use crate::{
@@ -92,9 +93,13 @@ async fn main() -> anyhow::Result<()> {
                 description = built_info::PKG_DESCRIPTION
             );
 
+            // Watches for the SIGTERM signal and sends a signal to all receivers, which gracefully
+            // shuts down all concurrent tasks below (EoS checker, controller, webhook server).
+            let sigterm_watcher = SignalWatcher::sigterm()?;
+
             let eos_checker =
                 EndOfSupportChecker::new(built_info::BUILT_TIME_UTC, maintenance.end_of_support)?
-                    .run()
+                    .run(sigterm_watcher.handle())
                     .map(anyhow::Ok);
 
             let product_config = product_config.load(&[
@@ -132,7 +137,6 @@ async fn main() -> anyhow::Result<()> {
                     watch_namespace.get_api::<StatefulSet>(&client),
                     watcher::Config::default(),
                 )
-                .shutdown_on_signal()
                 .watches(
                     client
                         .get_api::<DeserializeGuard<auth_core::v1alpha1::AuthenticationClass>>(&()),
@@ -160,6 +164,7 @@ async fn main() -> anyhow::Result<()> {
                             .map(|airflow| ObjectRef::from_obj(&*airflow))
                     },
                 )
+                .graceful_shutdown_on(sigterm_watcher.handle())
                 .run(
                     airflow_controller::reconcile_airflow,
                     airflow_controller::error_policy,
@@ -195,7 +200,7 @@ async fn main() -> anyhow::Result<()> {
             .await?;
 
             let webhook_server = webhook_server
-                .run()
+                .run(sigterm_watcher.handle())
                 .map_err(|err| anyhow!(err).context("failed to run webhook server"));
 
             futures::try_join!(airflow_controller, webhook_server, eos_checker)?;